Ensure return type is float in measure_time function

+++ python_async_function/2-measure_runtime.py
@@ -12,4 +12,4 @@ async def measure_time(n: int, max_delay: float) -> float:
     await wait_n(n, max_delay)
     end = time.time()
     total_time = end - start
-    return total_time / n
+    return float(total_time / n)
